@@ -1,1 +1,8 @@
-module.exports = require('@oclif/command')
+import cmd from '@oclif/command'
+import updateNotifier from 'update-notifier'
+
+import pkg from '../package.json'
+
+updateNotifier({ pkg }).notify()
+
+module.exports = cmd
