@@ -691,14 +691,14 @@ describe('api-bzz-node', () => {
 
   it('pins when uploading', async () => {
     const pinsBefore = await bzz.pins()
-    await Promise.all(pinsBefore.map(p => bzz.unpin(p.hash)))
+    await Promise.all(pinsBefore.map(p => bzz.unpin(p.address)))
 
     const hash = await bzz.uploadFile(uploadContent, { pin: true })
     const pins = await bzz.pins()
     expect(pins.length).toBe(1)
     expect(pins[0]).toEqual({
-      hash,
-      pinCounter: 1,
+      address: hash,
+      counter: 1,
       raw: true,
       size: uploadContent.length,
     })
@@ -710,7 +710,7 @@ describe('api-bzz-node', () => {
 
   it('pins an already uploaded raw file', async () => {
     const pinsBefore = await bzz.pins()
-    await Promise.all(pinsBefore.map(p => bzz.unpin(p.hash)))
+    await Promise.all(pinsBefore.map(p => bzz.unpin(p.address)))
 
     const hash = await bzz.uploadFile(uploadContent)
     await bzz.pin(hash, { raw: true })
@@ -718,8 +718,8 @@ describe('api-bzz-node', () => {
     const pins = await bzz.pins()
     expect(pins.length).toBe(1)
     expect(pins[0]).toEqual({
-      hash,
-      pinCounter: 1,
+      address: hash,
+      counter: 1,
       raw: true,
       size: uploadContent.length,
     })
@@ -731,7 +731,7 @@ describe('api-bzz-node', () => {
 
   it('pins a manifest', async () => {
     const pinsBefore = await bzz.pins()
-    await Promise.all(pinsBefore.map(p => bzz.unpin(p.hash)))
+    await Promise.all(pinsBefore.map(p => bzz.unpin(p.address)))
 
     const hash = await bzz.uploadFile(uploadContent, {
       contentType: 'text/plain',
@@ -740,11 +740,36 @@ describe('api-bzz-node', () => {
 
     const pins = await bzz.pins()
     expect(pins.length).toBe(1)
-    expect(pins[0].hash).toBe(hash)
+    expect(pins[0].address).toBe(hash)
     expect(pins[0].raw).toBe(false)
 
     await bzz.unpin(hash)
     const pinsAfter = await bzz.pins()
     expect(pinsAfter.length).toBe(0)
+  })
+
+  it('gets an upload tag', async () => {
+    const hash = await bzz.uploadFile(uploadContent, {
+      contentType: 'text/plain',
+    })
+    const tag = await bzz.getTag(hash)
+    expect(tag.address).toBe(hash)
+  })
+
+  it('polls an upload tag', async done => {
+    const hash = await bzz.uploadFile(uploadContent, {
+      contentType: 'text/plain',
+    })
+
+    let count = 0
+    const sub = bzz.pollTag(hash, { interval: 300 }).subscribe({
+      next: tag => {
+        expect(tag.address).toBe(hash)
+        if (count++ === 3) {
+          sub.unsubscribe()
+          done()
+        }
+      },
+    })
   })
 })
