@@ -89,17 +89,20 @@ export interface UploadOptions extends FileOptions {
 export interface PollOptions extends FetchOptions {
   interval: number // in milliseconds
   immediate?: boolean // defaults to true
+}
+
+export interface PollFeedOptions extends PollOptions {
   whenEmpty?: 'accept' | 'ignore' | 'error' // defaults to 'accept'
   trigger?: Observable<void>
 }
 
-export interface PollContentHashOptions extends PollOptions {
+export interface PollFeedContentHashOptions extends PollFeedOptions {
   changedOnly?: boolean
 }
 
-export interface PollContentOptions
+export interface PollFeedContentOptions
   extends DownloadOptions,
-    PollContentHashOptions {}
+    PollFeedContentHashOptions {}
 
 export interface PinOptions extends FetchOptions {
   download?: boolean
@@ -107,10 +110,23 @@ export interface PinOptions extends FetchOptions {
 }
 
 export interface PinnedFile {
-  hash: string
-  pinCounter: number
+  address: string
+  counter: number
   raw: boolean
   size: number
+}
+
+export interface Tag {
+  uid: number
+  name: string
+  address: string
+  total: number
+  split: number
+  seen: number
+  stored: number
+  sent: number
+  synced: number
+  startedAt: Date
 }
 
 export interface FeedParams {
