@@ -1,7 +1,7 @@
 /// <reference types="node" />
 import { hexInput, hexValue } from '@erebos/hex';
 import { Observable } from 'rxjs';
-import { BaseResponse, RequestInit, Fetch, BzzConfig, BzzMode, DirectoryData, DownloadOptions, FeedMetadata, FeedParams, FeedUpdateParams, FetchOptions, ListResult, PinOptions, PinnedFile, PollOptions, PollContentHashOptions, PollContentOptions, SignBytesFunc, UploadOptions } from './types';
+import { BaseResponse, RequestInit, Fetch, BzzConfig, BzzMode, DirectoryData, DownloadOptions, FeedMetadata, FeedParams, FeedUpdateParams, FetchOptions, ListResult, PinOptions, PinnedFile, PollOptions, PollFeedOptions, PollFeedContentHashOptions, PollFeedContentOptions, SignBytesFunc, Tag, UploadOptions } from './types';
 export * from './feed';
 export * from './types';
 export declare const BZZ_MODE_PROTOCOLS: {
@@ -10,6 +10,7 @@ export declare const BZZ_MODE_PROTOCOLS: {
     immutable: string;
     pin: string;
     raw: string;
+    tag: string;
 };
 export declare function getModeProtocol(mode?: BzzMode): string;
 export declare class HTTPError extends Error {
@@ -32,6 +33,7 @@ export declare class BaseBzz<Response extends BaseResponse> {
     getDownloadURL(hash: string, options?: DownloadOptions, raw?: boolean): string;
     getUploadURL(options?: UploadOptions, raw?: boolean): string;
     getFeedURL(hashOrParams: string | FeedParams | FeedUpdateParams, flag?: 'meta'): string;
+    getPinURL(hash?: string, raw?: boolean): string;
     hash(domain: string, options?: FetchOptions): Promise<hexValue>;
     list(hash: string, options?: DownloadOptions): Promise<ListResult>;
     download(hash: string, options?: DownloadOptions): Promise<Response>;
@@ -45,9 +47,9 @@ export declare class BaseBzz<Response extends BaseResponse> {
     getFeedChunk(hashOrParams: string | FeedParams, options?: FetchOptions): Promise<Response>;
     getFeedContentHash(hashOrParams: string | FeedParams, options?: FetchOptions): Promise<string>;
     getFeedContent(hashOrParams: string | FeedParams, options?: DownloadOptions): Promise<Response>;
-    pollFeedChunk(hashOrParams: string | FeedParams, options: PollOptions): Observable<Response>;
-    pollFeedContentHash(hashOrParams: string | FeedParams, options: PollContentHashOptions): Observable<string | null>;
-    pollFeedContent(hashOrParams: string | FeedParams, options: PollContentOptions): Observable<Response | null>;
+    pollFeedChunk(hashOrParams: string | FeedParams, options: PollFeedOptions): Observable<Response>;
+    pollFeedContentHash(hashOrParams: string | FeedParams, options: PollFeedContentHashOptions): Observable<string | null>;
+    pollFeedContent(hashOrParams: string | FeedParams, options: PollFeedContentOptions): Observable<Response | null>;
     postSignedFeedChunk(params: FeedUpdateParams, body: Buffer, options?: FetchOptions): Promise<Response>;
     postFeedChunk(meta: FeedMetadata, data: hexInput, options?: FetchOptions, signParams?: any): Promise<Response>;
     setFeedChunk(hashOrParams: string | FeedParams, data: hexInput, options?: FetchOptions, signParams?: any): Promise<Response>;
@@ -56,4 +58,6 @@ export declare class BaseBzz<Response extends BaseResponse> {
     pin(hash: string, options?: PinOptions): Promise<void>;
     unpin(hash: string, options?: FetchOptions): Promise<void>;
     pins(options?: FetchOptions): Promise<Array<PinnedFile>>;
+    getTag(hash: string, options?: FetchOptions): Promise<Tag>;
+    pollTag(hash: string, options: PollOptions): Observable<Tag>;
 }
