@@ -71,23 +71,37 @@ export interface UploadOptions extends FileOptions {
 export interface PollOptions extends FetchOptions {
     interval: number;
     immediate?: boolean;
+}
+export interface PollFeedOptions extends PollOptions {
     whenEmpty?: 'accept' | 'ignore' | 'error';
     trigger?: Observable<void>;
 }
-export interface PollContentHashOptions extends PollOptions {
+export interface PollFeedContentHashOptions extends PollFeedOptions {
     changedOnly?: boolean;
 }
-export interface PollContentOptions extends DownloadOptions, PollContentHashOptions {
+export interface PollFeedContentOptions extends DownloadOptions, PollFeedContentHashOptions {
 }
 export interface PinOptions extends FetchOptions {
     download?: boolean;
     raw?: boolean;
 }
 export interface PinnedFile {
-    hash: string;
-    pinCounter: number;
+    address: string;
+    counter: number;
     raw: boolean;
     size: number;
+}
+export interface Tag {
+    uid: number;
+    name: string;
+    address: string;
+    total: number;
+    split: number;
+    seen: number;
+    stored: number;
+    sent: number;
+    synced: number;
+    startedAt: Date;
 }
 export interface FeedParams {
     user: string;
