@@ -36,7 +36,7 @@ export default class BzzUploadCommand extends Command<Flags, Args> {
     }),
   }
 
-  public async run() {
+  public async run(): Promise<void> {
     this.spinner.start(`Uploading contents from: ${this.args.path}`)
     try {
       const hash = await this.client.bzz.uploadFrom(
