@@ -3,7 +3,7 @@ import { Command, DefaultFlags } from '../../Command'
 export default class PssAddressCommand extends Command<DefaultFlags> {
   public static flags = Command.flags
 
-  public async run() {
+  public async run(): Promise<void> {
     this.spinner.start('Retrieving PSS base address')
     try {
       const address = await this.client.pss.baseAddr()
