@@ -3,7 +3,7 @@ import { Command, DefaultFlags } from '../../Command'
 export default class PssKeyCommand extends Command<DefaultFlags> {
   public static flags = Command.flags
 
-  public async run() {
+  public async run(): Promise<void> {
     this.spinner.start('Retrieving PSS public key')
     try {
       const key = await this.client.pss.getPublicKey()
